+++ crates/gpui/src/input.rs
@@ -1,15 +1,12 @@
 mod bidi;
-mod bindings;
-mod input_handler;
-mod input_state;
+/// Input keybinding configuration & actions that can be bound (`Backspace`, `Copy`, etc.).
+///
+/// Explicitly not exported using `pub use bindings::*` to avoid namespace pollution.
+pub mod bindings;
+mod handler;
+mod state;
 
 pub use bidi::{TextDirection, detect_base_direction};
 pub use bindings::{INPUT_CONTEXT, InputBindings, bind_input_keys};
-pub use input_handler::*;
-pub use input_state::{InputLineLayout, InputState};
-// todo: move to keymap
-pub use input_state::{
-    Backspace, Copy, Cut, Delete, Down, End, Enter, Home, Left, MoveToBeginning, MoveToEnd, Paste,
-    Redo, Right, SelectAll, SelectDown, SelectLeft, SelectRight, SelectToBeginning, SelectToEnd,
-    SelectUp, SelectWordLeft, SelectWordRight, Tab, Undo, Up, WordLeft, WordRight,
-};
+pub use handler::*;
+pub use state::{InputLineLayout, InputState};
