+++ crates/gpui/src/input/bindings.rs
@@ -1,10 +1,66 @@
-use crate::{App, KeyBinding};
+use crate::{App, KeyBinding, actions};
 
-use super::{
-    Backspace, Copy, Cut, Delete, Down, End, Enter, Home, Left, MoveToBeginning, MoveToEnd, Paste,
-    Redo, Right, SelectAll, SelectDown, SelectLeft, SelectRight, SelectToBeginning, SelectToEnd,
-    SelectUp, SelectWordLeft, SelectWordRight, Tab, Undo, Up, WordLeft, WordRight,
-};
+actions!(
+    input,
+    [
+        /// Delete the character before the cursor.
+        Backspace,
+        /// Delete the character after the cursor.
+        Delete,
+        /// Insert a tab character at the cursor position.
+        Tab,
+        /// Move the cursor one character to the left.
+        Left,
+        /// Move the cursor one character to the right.
+        Right,
+        /// Move the cursor up one visual line.
+        Up,
+        /// Move the cursor down one visual line.
+        Down,
+        /// Extend selection one character to the left.
+        SelectLeft,
+        /// Extend selection one character to the right.
+        SelectRight,
+        /// Extend selection up one visual line.
+        SelectUp,
+        /// Extend selection down one visual line.
+        SelectDown,
+        /// Select all text content.
+        SelectAll,
+        /// Move cursor to the start of the current line.
+        Home,
+        /// Move cursor to the end of the current line.
+        End,
+        /// Extend selection to the beginning of the content.
+        SelectToBeginning,
+        /// Extend selection to the end of the content.
+        SelectToEnd,
+        /// Move cursor to the beginning of the content.
+        MoveToBeginning,
+        /// Move cursor to the end of the content.
+        MoveToEnd,
+        /// Paste from clipboard at the cursor position.
+        Paste,
+        /// Cut selected text to clipboard.
+        Cut,
+        /// Copy selected text to clipboard.
+        Copy,
+        /// Insert a newline at the cursor position.
+        Enter,
+        /// Move cursor one word to the left.
+        WordLeft,
+        /// Move cursor one word to the right.
+        WordRight,
+        /// Extend selection one word to the left.
+        SelectWordLeft,
+        /// Extend selection one word to the right.
+        SelectWordRight,
+        /// Undo the last edit.
+        Undo,
+        /// Redo the last undone edit.
+        Redo,
+    ]
+);
 
 /// The key context used for input element keybindings.
 pub const INPUT_CONTEXT: &str = "Input";
